Fix admin menu collapsing
