+++ includes/post-support_ticket.php
@@ -92,36 +92,6 @@ function remove_quick_edit_link( $actions, $post ) {
 add_filter( 'post_row_actions', 'ucare\remove_quick_edit_link', 10, 2 );
 
 
-function force_menu_expand( $file ) {
-
-    $screen = get_current_screen();
-
-    if( $screen && $screen->taxonomy == 'ticket_category' ) {
-        return 'ucare_support';
-    }
-
-    return $file;
-
-}
-
-add_filter( 'parent_file', 'ucare\force_menu_expand' );
-
-
-function menu_highlight_categories( $file ) {
-
-    $screen = get_current_screen();
-
-    if( $screen && $screen->taxonomy == 'ticket_category' ) {
-        return 'edit-tags.php?post_type=support_ticket&taxonomy=ticket_category';
-    }
-
-    return $file;
-
-}
-
-add_filter( 'submenu_file', 'ucare\menu_highlight_categories' );
-
-
 function tickets_table_sortable_columns( $columns ) {
 
     $sortable = array(
@@ -311,7 +281,7 @@ add_filter( 'parse_query', 'ucare\filter_tickets_table' );
 
 function ticket_meta_boxes() {
 
-    //TODO Refactor these into a single metabox
+    //TODO Remove this and replace with regular metaboxes
     $support_metabox = new FormMetaBox(
         array(
             'id'        => 'ticket_support_meta',
@@ -343,3 +313,78 @@ function ticket_meta_boxes() {
 add_action( 'admin_init', 'ucare\ticket_meta_boxes' );
 
 
+function ticket_properties_updated( $null, $id, $key, $value ) {
+
+    global $wpdb;
+
+    if( get_post_type( $id ) == 'support_ticket' && $key == 'status' ) {
+
+        $q = "UPDATE {$wpdb->posts}
+              SET post_modified = %s, post_modified_gmt = %s
+              WHERE ID = %d ";
+
+        $q = $wpdb->prepare( $q, array( current_time( 'mysql' ), current_time( 'mysql', 1 ), $id ) );
+
+        $wpdb->query( $q );
+
+        delete_post_meta( $id, 'stale' );
+
+        if( $value == 'closed' ) {
+
+            update_post_meta( $id, 'closed_date', current_time( 'mysql' ) );
+            update_post_meta( $id, 'closed_by', wp_get_current_user()->ID );
+
+        }
+
+    }
+
+}
+
+add_action( 'update_post_metadata', 'ucare\ticket_properties_updated', 10, 4 );
+
+
+function set_default_ticket_meta( $post_id, $post, $update ) {
+
+    $defaults = array(
+        'priority' => 0
+    );
+
+    if( !$update ) {
+
+        foreach( $defaults as $key => $value ) {
+            add_post_meta( $post_id, $key, $value );
+        }
+
+    }
+
+}
+
+add_action( 'wp_insert_post', 'ucare\set_default_ticket_meta', 10, 3 );
+
+
+function get_recent_tickets( $args = array() ) {
+
+    $defaults = array(
+        'author'  => '',
+        'after'   => 'now',
+        'before'  => '30 days ago',
+        'exclude' => array(),
+        'limit'   => -1
+    );
+
+    $args = wp_parse_args( $args, $defaults );
+
+
+    $q = array(
+        'post_type'      => 'support_ticket',
+        'post_status'    => 'publish',
+        'author'         => $args['author'],
+        'after'          => $args['after'],
+        'before'         => $args['before'],
+        'post__not_in'   => $args['exclude'],
+        'posts_per_page' => $args['limit'] ?: -1
+    );
+
+    return new \WP_Query( $q );
+
+}
